+++ org.hl7.fhir.report/pom.xml
@@ -70,7 +70,7 @@
                             <footer>Code Coverage Report for FHIR core libs ${project.version}</footer>
                             <includes>
                                 <!-- Analyze class files only to exclude shaded agent JAR from report -->
-                                <include>**/*.class</include>
+                                <include>org/hl7/fhir/**/*.class</include>
                             </includes>
                         </configuration>
                     </execution>
@@ -84,6 +84,14 @@
                     <!-- The report project pulls in resources from all the other packages, which results in some duplicate
                      classes being pulled in. This will need to be fixed eventually, but for now, we can just disable the
                      check on this package. It's only purpose is to aggregate the reports.-->
+                    <skip>true</skip>
+                </configuration>
+            </plugin>
+            <plugin>
+                <groupId>org.owasp</groupId>
+                <artifactId>dependency-check-maven</artifactId>
+                <configuration>
+                    <!-- -->
                     <skip>true</skip>
                 </configuration>
             </plugin>
