+++ org.hl7.fhir.validation.cli/src/test/java/org/hl7/fhir/validation/cli/ValidatorCliTests.java
@@ -1,4 +1,4 @@
-package org.hl7.fhir.validation;
+package org.hl7.fhir.validation.cli;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.fail;
@@ -13,6 +13,7 @@ import static org.mockito.Mockito.when;
 import java.util.List;
 
 import org.hl7.fhir.utilities.TimeTracker;
+import org.hl7.fhir.validation.ValidationEngine;
 import org.hl7.fhir.validation.cli.model.CliContext;
 import org.hl7.fhir.validation.cli.services.ValidationService;
 import org.hl7.fhir.validation.cli.services.ValidatorWatchMode;
